+++ packages/sdk/src/Portfolio.ts
@@ -1,23 +1,25 @@
 import * as Abis from "@enzymefinance/abis";
-import type { Address, PublicClient } from "viem";
-import * as Assets from "./Assets.js";
-
+import type { Address, Hex, PublicClient } from "viem";
+import * as Assets from "./Asset.js";
 import { Assertion, Viem } from "./Utils.js";
-
 export * as Integrations from "./Portfolio/Integrations.js";
 export * as VoteDelegation from "./Portfolio/VoteDelegation.js";
 
 export {
   Action as ExternalPositionAction,
+  call as callExternalPosition,
   callEncode as callExternalPositionEncode,
   callDecode as callExternalPositionDecode,
   type CallArgs as ExternalPositionCallArgs,
+  create as createExternalPosition,
   createEncode as createExternalPositionEncode,
   createDecode as createExternalPositionDecode,
   type CreateArgs as CreateExternalPositionArgs,
+  remove as removeExternalPosition,
   removeEncode as removeExternalPositionEncode,
   removeDecode as removeExternalPositionDecode,
   type RemoveArgs as RemoveExternalPositionArgs,
+  reactivate as reactivateExternalPosition,
   reactivateEncode as reactivateExternalPositionEncode,
   reactivateDecode as reactivateExternalPositionDecode,
   type ReactivateArgs as ReactivateExternalPositionArgs,
@@ -26,16 +28,35 @@ export {
 export {
   Action as IntegrationAdapterAction,
   Selector as IntegrationAdapterSelector,
+  call as callIntegrationAdapter,
   callEncode as callIntegrationAdapterEncode,
   callDecode as callIntegrationAdapterDecode,
   type CallArgs as CallIntegrationAdapterArgs,
+  addTrackedAssets,
   addTrackedAssetsEncode,
   addTracketAssetsDecode,
   type AddTracketAssetsArgs,
+  removeTracketAssets,
   removeTrackedAssetsEncode,
   removeTrackedAssetsDecode,
   type RemoveTrackedAssetsArgs,
 } from "./_internal/IntegrationManager.js";
+
+export type VaultCallOnContractParams = {
+  comptrollerProxy: Address;
+  contract: Address;
+  selector: Hex;
+  encodedArgs: Hex;
+};
+
+export function vaultCallOnContract(args: VaultCallOnContractParams) {
+  return new Viem.PopulatedTransaction({
+    abi: Abis.IComptrollerLib,
+    functionName: "vaultCallOnContract",
+    address: args.comptrollerProxy,
+    args: [args.contract, args.selector, args.encodedArgs],
+  });
+}
 
 export async function getPortfolio(
   client: PublicClient,