+++ packages/sdk/src/Portfolio/Integrations/Erc4626.ts
@@ -6,41 +6,44 @@ import * as IntegrationManager from "../../_internal/IntegrationManager.js";
 // LEND
 //--------------------------------------------------------------------------------------------
 
-const lendSelector = "0x099f7515"; // lend(address,bytes,bytes)
-export const lend = IntegrationManager.makeUse(lendSelector, lendEncode);
+export const lend = IntegrationManager.makeUse(IntegrationManager.Selector.Lend, lendEncode);
 
 const lendEncoding = [
   {
-    name: "tokenAddress",
+    name: "erc4626Vault",
     type: "address",
   },
   {
-    name: "outgoingAssetAmount",
+    name: "underlyingAssetAmount",
     type: "uint256",
   },
   {
-    name: "minIncomingAmount",
+    name: "minIncomingSharesAmount",
     type: "uint256",
   },
 ] as const;
 
 export type LendArgs = {
-  tokenAddress: Address;
-  outgoingAssetAmount: bigint;
-  minIncomingAmount: bigint;
+  erc4626Vault: Address;
+  underlyingAssetAmount: bigint;
+  minIncomingSharesAmount: bigint;
 };
 
 export function lendEncode(args: LendArgs): Hex {
-  return encodeAbiParameters(lendEncoding, [args.tokenAddress, args.outgoingAssetAmount, args.minIncomingAmount]);
+  return encodeAbiParameters(lendEncoding, [
+    args.erc4626Vault,
+    args.underlyingAssetAmount,
+    args.minIncomingSharesAmount,
+  ]);
 }
 
 export function lendDecode(encoded: Hex): LendArgs {
-  const [tokenAddress, outgoingAssetAmount, minIncomingAmount] = decodeAbiParameters(lendEncoding, encoded);
+  const [erc4626Vault, underlyingAssetAmount, minIncomingSharesAmount] = decodeAbiParameters(lendEncoding, encoded);
 
   return {
-    tokenAddress,
-    outgoingAssetAmount,
-    minIncomingAmount,
+    erc4626Vault,
+    underlyingAssetAmount,
+    minIncomingSharesAmount,
   };
 }
 
@@ -48,38 +51,75 @@ export function lendDecode(encoded: Hex): LendArgs {
 // REDEEM
 //--------------------------------------------------------------------------------------------
 
-const redeemSelector = "0xc29fa9dd"; // redeem(address,bytes,bytes)
-export const redeem = IntegrationManager.makeUse(redeemSelector, redeemEncode);
+export const redeem = IntegrationManager.makeUse(IntegrationManager.Selector.Redeem, redeemEncode);
 
 const redeemEncoding = [
   {
-    name: "tokenAddress",
+    name: "erc4626Vault",
     type: "address",
   },
   {
-    name: "outgoingAssetAmount",
+    name: "sharesAmount",
     type: "uint256",
   },
   {
-    name: "minIncomingAmount",
+    name: "minIncomingUnderlyingAssetAmount",
     type: "uint256",
   },
 ] as const;
 
 export type RedeemArgs = {
-  tokenAddress: Address;
-  outgoingAssetAmount: bigint;
-  minIncomingAmount: bigint;
+  erc4626Vault: Address;
+  sharesAmount: bigint;
+  minIncomingUnderlyingAssetAmount: bigint;
 };
 
 export function redeemEncode(args: RedeemArgs): Hex {
-  return encodeAbiParameters(redeemEncoding, [args.tokenAddress, args.outgoingAssetAmount, args.minIncomingAmount]);
+  return encodeAbiParameters(redeemEncoding, [
+    args.erc4626Vault,
+    args.sharesAmount,
+    args.minIncomingUnderlyingAssetAmount,
+  ]);
 }
 
 export function redeemDecode(encoded: Hex): RedeemArgs {
-  const [tokenAddress, outgoingAssetAmount, minIncomingAmount] = decodeAbiParameters(redeemEncoding, encoded);
+  const [erc4626Vault, sharesAmount, minIncomingUnderlyingAssetAmount] = decodeAbiParameters(redeemEncoding, encoded);
 
-  return { tokenAddress, outgoingAssetAmount, minIncomingAmount };
+  return { erc4626Vault, sharesAmount, minIncomingUnderlyingAssetAmount };
+}
+
+//--------------------------------------------------------------------------------------------
+// READ FUNCTIONS
+//--------------------------------------------------------------------------------------------
+
+export async function convertToAssets(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    erc4626Vault: Address;
+    sharesAmount: bigint;
+  }>,
+) {
+  return Viem.readContract(client, args, {
+    abi: parseAbi(["function convertToAssets(uint256 sharesAmount) view returns (uint256 assetAmount)"]),
+    functionName: "convertToAssets",
+    address: args.erc4626Vault,
+    args: [args.sharesAmount],
+  });
+}
+
+export async function convertToShares(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    erc4626Vault: Address;
+    assetAmount: bigint;
+  }>,
+) {
+  return Viem.readContract(client, args, {
+    abi: parseAbi(["function convertToShares(uint256 assetAmount) view returns (uint256 sharesAmount)"]),
+    functionName: "convertToShares",
+    address: args.erc4626Vault,
+    args: [args.assetAmount],
+  });
 }
 
 //--------------------------------------------------------------------------------------------