+++ packages/sdk/src/Portfolio/Integrations/UniswapV3.ts
@@ -7,8 +7,7 @@ import * as IntegrationManager from "../../_internal/IntegrationManager.js";
 // TAKE ORDER
 //--------------------------------------------------------------------------------------------
 
-const takeOrderSelector = "0x03e38a2b"; // takeOrder(address,bytes,bytes)
-export const takeOrder = IntegrationManager.makeUse(takeOrderSelector, takeOrderEncode);
+export const takeOrder = IntegrationManager.makeUse(IntegrationManager.Selector.TakeOrder, takeOrderEncode);
 
 const takeOrderEncoding = [
   {
@@ -63,7 +62,7 @@ export function takeOrderDecode(encoded: Hex): TakeOrderArgs {
 // EXTERNAL POSITION
 //--------------------------------------------------------------------------------------------
 
-export type Action = typeof Action[keyof typeof Action];
+export type Action = (typeof Action)[keyof typeof Action];
 export const Action = {
   Mint: 0n,
   AddLiquidity: 1n,