+++ packages/sdk/vite.config.ts
@@ -1,5 +1,4 @@
-/// <reference types="vitest" />
-import path from "node:path";
+import * as path from "node:path";
 import aliases from "vite-tsconfig-paths";
 import { defineProject } from "vitest/config";